requirements.txt
transformers
torch
torchaudio
datasets
numpy
evaluate
rouge_score
bert_score
python-dotenv
huggingface_hub
tqdm 
jiwer
librosa
soundfile
peft
tensorboard
silero_vad
trl
bitsandbytes
unsloth